# gh-138213: Make `csv.reader` 1.4x faster