--- a/Doc/whatsnew/3.15.rst
+++ b/Doc/whatsnew/3.15.rst
@@ -427,7 +427,7 @@ csv
 ---
 
 * The :meth:`csv.reader` has been optimized, and is around 1.4x faster.
-  (Contributed by Maurycy Pawłowski-Wieroński in :gh:`XXX`.)
+  (Contributed by Maurycy Pawłowski-Wieroński in :gh:`138214`.)
 
 
 Deprecated