--- a/Doc/whatsnew/3.15.rst
+++ b/Doc/whatsnew/3.15.rst
@@ -426,7 +426,7 @@ Optimizations
 csv
 ---
 
-* The :meth:`csv.reader` has been optimized, and is around 1.4x faster.
+* The :func:`csv.reader` has been optimized, and is around 1.4x faster.
   (Contributed by Maurycy Pawłowski-Wieroński in :gh:`138214`.)