--- a/Misc/NEWS.d/next/Library/2025-08-28-02-41-14.gh-issue-138213.8m2OO9.rst
+++ b/Misc/NEWS.d/next/Library/2025-08-28-02-41-14.gh-issue-138213.8m2OO9.rst
@@ -1,1 +1,1 @@
-Speed up :class:`~csv.reader` by 1.4x.
+Speed up :class:`~csv.reader` by 2x.