The basic observation is that there's no need call the state machine while we're in a field (`IN_FIELD`, `IN_QUOTED_FIELD`).

Most characters are ordinary (ie: they're not delimiters, escapes, quotes etc.), so we can find the next interesting character and copy the whole slice.

This is my very first C change in cpython so I'm more than happy to pair with someone.

# Benchmark

<details>
<summary>The script:</summary>

```python
import csv
import io
import pyperf
import os

NUM_ROWS = (10_000, )
NUM_COLS = (5, )
FIELD_LENGTH = (30, )


def generate_csv_data(rows, cols, field_len):
    field = "x" * field_len
    row = ",".join([field] * cols)
    return os.linesep.join([row] * rows) + os.linesep


def benchmark_csv_reader(csv_data):
    for _ in csv.reader(io.StringIO(csv_data)):
        pass


runner = pyperf.Runner()

for rows in NUM_ROWS:
    for cols in NUM_COLS:
        for field_len in FIELD_LENGTH:
            csv_data = generate_csv_data(rows, cols, field_len)
            runner.bench_func(
                f"csv_reader({rows}, {cols}, {field_len})",
                benchmark_csv_reader,
                csv_data,
            )
```
</details>

The results:

```
Mean +- std dev: [bench_csv_reader.main] 8.16 ms +- 0.16 ms -> [bench_csv_reader.csv-read-chunks] 5.71 ms +- 0.08 ms: 1.43x faster
```

The environment:

```
% ./python -c "import sysconfig; print(sysconfig.get_config_var('CONFIG_ARGS'))"
'--enable-lto' '--with-optimizations'
```

`sudo ./python -m pyperf system tune` ensured.